Put the Apple Notes wording where OneNote keeps its own
describeFolderFailure sat in the 900-line importer, and reached for the
HTTP failure parser — status, code, nested error bodies, none of which
a SQLite error has — to get at a message extractErrorMessage already
returns. It moves to apple-notes/errors.ts, beside its OneNote twin,
and the importer stops importing request-failure at all.

The test now reaches the wording without loading the importer.
